moving operationTracking to parent level to share it between sendgrid and twilio
