+++ packages/destination-actions/src/destinations/engage/twilio/utils/IntegrationErrorWrapper.ts
@@ -1,5 +1,5 @@
 import { IntegrationError } from '@segment/actions-core'
-import { OperationErrorHandler, TryCatchFinallyContext, TryCatchFinallyHook } from '../operationTracking'
+import { OperationErrorHandler, TryCatchFinallyContext, TryCatchFinallyHook } from '../../operationTracking'
 import { TwilioApiError } from './types'
 
 export type IntegrationErrorWrapperContext<TContext extends TryCatchFinallyContext = TryCatchFinallyContext> =
